Seperate dataclass for easier import
