+++ src/py/flwr_experimental/baseline/command.py
@@ -17,7 +17,7 @@
 
 from typing import Optional
 
-from flwr_experimental.ops.cluster import Instance
+from flwr_experimental.ops.instance import Instance
 
 
 def install_wheel(wheel_remote_path: str) -> str:
