+++ src/py/flwr_experimental/baseline/tf_cifar/settings.py
@@ -21,7 +21,7 @@ from flwr_experimental.baseline.common import (
     sample_delay_factors,
 )
 from flwr_experimental.baseline.setting import Baseline, ClientSetting, ServerSetting
-from flwr_experimental.ops.cluster import Instance
+from flwr_experimental.ops.instance import Instance
 
 ROUNDS = 20
 MIN_NUM_CLIENTS = 80
